+++ SMath.cpp
