+++ SMath.h
